(1) Fixed and Eclipsed vs. junit bug in PropertyManagerTest. uggh. (2) Started working on cleaning up the public interface of Agent. In particular, making fewer internal components (trace, exploration, etc) public. Using Adaptables framework as a pseudo-friend mechanism to allow components to be private but still accessible to the spaghetti that lies beneath.
--HG--
extra : convert_revision : svn%3A849a83f0-50c2-4045-bdd1-8b1f698b4b18/trunk%40205

+++ jsoar/test/org/jsoar/util/properties/PropertyManagerTest.java
@@ -42,7 +42,7 @@ public class PropertyManagerTest
         PropertyManager pm = new PropertyManager();
         
         assertEquals("default", pm.get(KEY));
-        assertEquals(666, pm.get(KEY2));
+        assertEquals(666, pm.get(KEY2).intValue());
     }
     
     @Test
@@ -56,9 +56,9 @@ public class PropertyManagerTest
         assertEquals("default", pm.set(KEY, "new"));
         assertEquals("new", pm.get(KEY));
         
-        assertEquals(666, pm.get(KEY2));
-        assertEquals(666, pm.set(KEY2, 555));
-        assertEquals(555, pm.get(KEY2));
+        assertEquals(666, pm.get(KEY2).intValue());
+        assertEquals(666, pm.set(KEY2, 555).intValue());
+        assertEquals(555, pm.get(KEY2).intValue());
     }
     
     @Test
@@ -118,8 +118,8 @@ public class PropertyManagerTest
             {
                 assertNotNull(event);
                 assertEquals(KEY2, event.getKey());
-                assertEquals(666, event.getOldValue());
-                assertEquals(555, event.getNewValue());
+                assertEquals(666, event.getOldValue().intValue());
+                assertEquals(555, event.getNewValue().intValue());
                 key2Count[0]++;
             }});
         
@@ -127,9 +127,9 @@ public class PropertyManagerTest
         assertEquals("default", pm.set(KEY, "new"));
         assertEquals("new", pm.get(KEY));
         
-        assertEquals(666, pm.get(KEY2));
-        assertEquals(666, pm.set(KEY2, 555));
-        assertEquals(555, pm.get(KEY2));
+        assertEquals(666, pm.get(KEY2).intValue());
+        assertEquals(666, pm.set(KEY2, 555).intValue());
+        assertEquals(555, pm.get(KEY2).intValue());
         
         assertEquals(1, key1Count[0]);
         assertEquals(1, key2Count[0]);
